Revert "feat(auto_authn): modularize auth flows"
