--- pkgs/standards/auto_authn/auto_authn/v2/routers/flows/__init__.py
@@ -1,13 +1,0 @@
-"""Modular OAuth/OIDC flow routers."""
-
-from . import credentials, authorization, token, logout, refresh, introspection, common
-
-__all__ = [
-    "credentials",
-    "authorization",
-    "token",
-    "logout",
-    "refresh",
-    "introspection",
-    "common",
-]
