--- pkgs/standards/auto_authn/auto_authn/v2/routers/flows/refresh.py
@@ -1,22 +1,0 @@
-from __future__ import annotations
-
-from fastapi import APIRouter, HTTPException, Request, status
-from pydantic import BaseModel
-
-from .common import TokenPair, _jwt, require_tls
-
-router = APIRouter()
-
-
-class RefreshIn(BaseModel):
-    refresh_token: str
-
-
-@router.post("/token/refresh", response_model=TokenPair)
-async def refresh(body: RefreshIn, request: Request):
-    require_tls(request)
-    try:
-        access, refresh = _jwt.refresh(body.refresh_token)
-    except Exception:
-        raise HTTPException(status.HTTP_401_UNAUTHORIZED, "invalid refresh token")
-    return TokenPair(access_token=access, refresh_token=refresh)
